refctoring: make modules JS
