+++ index.html
@@ -98,6 +98,6 @@
   </template>
 
   <!--подключение скрипта-->
-  <script src="./scripts/index.js"></script>
+  <script type="module" src="./scr/index.js"></script>
 </body>
 </html>
